refactor: [박지영] notice 도메인 백엔드 board->boards.notice 로 리팩토링 [BYL-BACK-PROCESS-16]
- 기존의 board 였던 도메인을 notice 도메인으로 바꾸면서
board -> boards.notice 로 리팩토링 하였습니다
