+++ src/main/java/com/jgj/byl_process/domain/boards/notice/service/NoticeBoardService.java
@@ -1,7 +1,7 @@
-package com.jgj.byl_process.domain.board.service;
+package com.jgj.byl_process.domain.boards.notice.service;
 
-import com.jgj.byl_process.domain.board.controller.request.NoticeBoardRequest;
-import com.jgj.byl_process.domain.board.entity.NoticeBoard;
+import com.jgj.byl_process.domain.boards.notice.entity.NoticeBoard;
+import com.jgj.byl_process.domain.boards.notice.service.request.NoticeBoardRequest;
 
 import java.util.List;
